requirements.txt
mistralai==0.4.2
python-dotenv==1.0.0
streamlit>=1.28.0
